fix(fp8): hand the W8A8 MoE kernel its activation scale in the compute dtype
The W8A8 branch of Fp8MoEMethod.apply passed custom_moe_glu_w8a8 a fp32
per-(token, K-block) scale. The op's fake declares the output in
compute_dtype, but the compiler drops compute_dtype and binds the kernel's
__DTYPE__ (scales, accumulator and output) from the scale operand, so the
compiled kernel emitted fp32 hidden_states while the graph believed bf16.
Inside a graph the next norm hides it; under pipeline parallelism the last
MoE output of a stage is the graph output, so PP1 rejected the handoff
("previous stage sent hidden_states float32, this stage expects bfloat16").

Round the scale to compute_dtype first and quantize with the rounded value,
so the kernel runs in its documented bf16 mode and dequantizes with exactly
the scale the values were divided by.

+++ vllm_rbln/model_executor/layers/quantization/fp8.py
@@ -342,9 +342,12 @@ class Fp8MoEMethod(FusedMoEMethodBase):
             hs_shape = hidden_states.shape
             s_g = hidden_states.reshape(-1, in_block_size).to(torch.float32)
             amax = s_g.abs().amax(dim=-1, keepdim=True).clamp_min(1e-10)
-            scale = amax / finfo.max
+            # The compiled kernel binds its __DTYPE__ (and so its output dtype)
+            # from this scale, so it must be compute_dtype, not fp32. Round it
+            # first and quantize with the rounded value.
+            scale = (amax / finfo.max).to(compute_dtype)
             hs_q = (
-                (s_g / scale)
+                (s_g / scale.float())
                 .clamp(finfo.min, finfo.max)
                 .to(fp8_dtype)
                 .reshape(hs_shape)
